+++ CHANGELOG.md
@@ -22,6 +22,9 @@ To install release candidates run `[sudo] gem install cocoapods --pre`
   [Dimitris Koutsogiorgas](https://github.com/dnkoutso)
   [#11213](https://github.com/CocoaPods/CocoaPods/pull/11213)
 
+* Run post install hooks when "skip Pods.xcodeproj generation" option is set
+  [Elton Gao](https://github.com/gyfelton)
+  [#11073](https://github.com/CocoaPods/CocoaPods/pull/11073)
 
 ## 1.11.2 (2021-09-13)